+++ test/calc_duration_test.dart
@@ -22,14 +22,14 @@ void main() {
     test('bass bar 1, q 2 = 2', () async {
       final song = await getSong('duration0');
 
-      final int? duration = calcDuration(1, 2, 'bass', song.bars);
+      final double? duration = calcDuration(1, 2, 'bass', song.bars);
       expect(duration, equals(2));
     });
 
     test('duration1 (n = 2): bass bar 0, q 0 = null', () async {
       final song = await getSong('duration1');
 
-      final int? duration = calcDuration(0, 0, 'bass', song.bars);
+      final double? duration = calcDuration(0, 0, 'bass', song.bars);
       expect(duration, equals(null));
     });
   });
@@ -38,49 +38,49 @@ void main() {
     test('bass bar 0, q 0 = 3', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(0, 0, 'bass', song.bars);
+      final double? duration = calcDuration(0, 0, 'bass', song.bars);
       expect(duration, equals(3));
     });
 
     test('bass bar 0, q 3 = 3', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(0, 3, 'bass', song.bars);
+      final double? duration = calcDuration(0, 3, 'bass', song.bars);
       expect(duration, equals(3));
     });
 
     test('bass bar 1, q 2 = 2', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(1, 2, 'bass', song.bars);
+      final double? duration = calcDuration(1, 2, 'bass', song.bars);
       expect(duration, equals(2));
     });
 
     test('bass bar 2, q 0 = 1', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(2, 0, 'bass', song.bars);
+      final double? duration = calcDuration(2, 0, 'bass', song.bars);
       expect(duration, equals(1));
     });
 
     test('bass bar 2, q 1 = 2', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(2, 1, 'bass', song.bars);
+      final double? duration = calcDuration(2, 1, 'bass', song.bars);
       expect(duration, equals(2));
     });
 
     test('bass bar 2, q 3 = 3', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(2, 3, 'bass', song.bars);
+      final double? duration = calcDuration(2, 3, 'bass', song.bars);
       expect(duration, equals(3));
     });
 
     test('bass bar 3, q 2 = 7', () async {
       final song = await getSong('calc_duration_test');
 
-      final int? duration = calcDuration(3, 2, 'bass', song.bars);
+      final double? duration = calcDuration(3, 2, 'bass', song.bars);
       expect(duration, equals(null));
     });
   });
@@ -89,14 +89,14 @@ void main() {
     test('bass bar 8, q 0 = 4', () async {
       final song = await getSong('After All');
 
-      final int? duration = calcDuration(8, 0, 'bass', song.bars);
+      final double? duration = calcDuration(8, 0, 'bass', song.bars);
       expect(duration, equals(4));
     });
 
     test('bass bar 16, q 1 = 1', () async {
       final song = await getSong('After All');
 
-      final int? duration = calcDuration(16, 1, 'bass', song.bars);
+      final double? duration = calcDuration(16, 1, 'bass', song.bars);
       expect(duration, equals(1));
     });
   });
@@ -105,7 +105,7 @@ void main() {
     test('bass bar 0, q 0 = 4', () async {
       final song = await getSong('scheduler');
 
-      final int? duration = calcDuration(0, 0, 'bass', song.bars);
+      final double? duration = calcDuration(0, 0, 'bass', song.bars);
       expect(duration, equals(null));
     });
   });
